Rewrite component version tracking
Make the component version atomic so that we can update the version
from multiple threads.

Rewrite ``ComponentSlice``, ``ComponentSliceMut`` and component
views to make slice's fields private.

Create the component's next version value in ``ComponentSliceMut``
constructor and store it in this structure.

Update the component version when user gets the reference to
the component data by dereferencing or by calling ``into_slice()``.